+++ src/lib.rs
@@ -183,6 +183,7 @@ pub mod mip;
 pub mod mip_tables;
 pub mod mvd_coding_enc;
 pub mod nal;
+pub mod non_merge_inter_pre_residual_dec;
 pub mod non_merge_inter_pre_residual_enc;
 pub mod non_merge_mvp_syntax_enc;
 pub mod opi;